AANSELMO - Implementado evento de seleção do filtro para atualizar a página com os novos dados.

+++ ExpensesControl/Areas/Access/Controllers/PasswordController.cs
@@ -75,7 +75,7 @@ namespace ExpensesControl.Areas.Access.Controllers
         }
 
         [HttpPost]
-        public IActionResult GenerateNewPassword([FromForm] PasswordReset passwordReset, int id)
+        public IActionResult GenerateNewPassword([FromForm] PasswordReset passwordReset, long id)
         {
             if (ModelState.IsValid)
             {
