+++ ExpensesControl/Repositories/Contracts/IUserRepository.cs
@@ -11,15 +11,15 @@ namespace ExpensesControl.Repositories
     public interface IUserRepository
     {
         public void Create(User user);
-        public User Read(int id);
+        public User Read(long id);
         public IPagedList<User> ReadAllUsers(int? page, SearchTypeUser searchTypeUser, string nameSearchField, OrdinationType ordination);
         public User ReadByEmail(string email);
         public void Update(User user);
         public void UpdateRegistrationData(User user);
         public void UpdatePassword(User user);
-        public void Delete(int id);
+        public void Delete(long id);
         public User Login(User user);
-        public void UpdateUserStatus(int id, UserStatus status);
+        public void UpdateUserStatus(long id, UserStatus status);
         public void UpdateEmail(User user);
     }
 }
